Additional Refactoring, Fix ShaderParameter Construction

+++ src/webgpu_utils.cpp
@@ -18,7 +18,7 @@ uint32_t ceilToNextMultiple(uint32_t value, uint32_t step) {
     return step * divide_and_ceil;
 }
 
-BufferHolder createBuffer(Device &device, uint size, WGPUBufferUsageFlags usage, bool mappedAtCreation) {
+BufferHolder createBuffer(Device &device, size_t size, WGPUBufferUsageFlags usage, bool mappedAtCreation) {
     BufferDescriptor bufferDesc;
     bufferDesc.size = size;
     bufferDesc.usage = usage;
